+++ src/ray/core_worker/object_interface.cc
@@ -42,7 +42,7 @@ Status CoreWorkerObjectInterface::Get(const std::vector<ObjectID> &ids,
   bool was_blocked = false;
 
   std::unordered_map<ObjectID, int> unready;
-  for (int i = 0; i < ids.size(); i++) {
+  for (size_t i = 0; i < ids.size(); i++) {
     unready.insert({ids[i], i});
   }
 
@@ -90,7 +90,7 @@ Status CoreWorkerObjectInterface::Get(const std::vector<ObjectID> &ids,
           core_worker_.store_client_.Get(plasma_ids, get_timeout, &object_buffers);
     }
 
-    for (int i = 0; i < object_buffers.size(); i++) {
+    for (size_t i = 0; i < object_buffers.size(); i++) {
       if (object_buffers[i].data != nullptr) {
         const auto &object_id = unready_ids[i];
         (*results)[unready[object_id]] =
@@ -126,7 +126,7 @@ Status CoreWorkerObjectInterface::Wait(const std::vector<ObjectID> &object_ids,
   // TODO: change RayletClient::Wait() to return a bit set, so that we don't need
   // to do this translation.
   (*results).resize(object_ids.size());
-  for (int i = 0; i < object_ids.size(); i++) {
+  for (size_t i = 0; i < object_ids.size(); i++) {
     (*results)[i] = ready_ids.count(object_ids[i]) > 0;
   }
